Fix seed script and refactor middleware
