+++ portals/developer-portal/src/routes/api/handlers/keyManagersHandler.js
@@ -20,7 +20,7 @@
 /*
  * Tag: Key Managers
  */
-const keyManagerService = require('../../services/keyManagerService');
+const keyManagerService = require('../../../services/keyManagerService');
 
 module.exports = {
     createKeyManager: keyManagerService.createKeyManager,
