+++ portals/developer-portal/src/routes/api/handlers/labelsHandler.js
@@ -20,7 +20,7 @@
 /*
  * Tag: Labels
  */
-const apiMetadataService = require('../../services/apiMetadataService');
+const apiMetadataService = require('../../../services/apiMetadataService');
 
 module.exports = {
     createLabels: apiMetadataService.createLabels,
