+++ portals/developer-portal/src/routes/api/handlers/subscriptionPoliciesHandler.js
@@ -20,7 +20,7 @@
 /*
  * Tag: Subscription Policies
  */
-const apiMetadataService = require('../../services/apiMetadataService');
+const apiMetadataService = require('../../../services/apiMetadataService');
 
 module.exports = {
     listSubscriptionPolicies: apiMetadataService.listSubscriptionPolicies,
